Updated createEventActivity and CreateEventOptionsActivity to be able to create an event
