+++ Project_v2/app/src/main/java/com/example/projectv2/View/EventLandingPageActivity.java
@@ -1,4 +1,4 @@
-package com.example.dummy.View;
+package com.example.projectv2.View;
 
 import android.content.Intent;
 import android.net.Uri;
@@ -8,66 +8,8 @@ import android.widget.ImageButton;
 import android.widget.ImageView;
 import android.widget.TextView;
 import androidx.appcompat.app.AppCompatActivity;
-import com.example.dummy.R;
 
 public class EventLandingPageActivity extends AppCompatActivity {
 
-    private ImageView eventImageView;
-    private TextView eventNameView, eventDetailsView, eventRulesView, eventDeadlineView, eventPriceView, eventCountdownView;
-
-    @Override
-    protected void onCreate(Bundle savedInstanceState) {
-        super.onCreate(savedInstanceState);
-        setContentView(R.layout.event_landing_page_organiser);
-
-        // Initialize views
-        eventImageView = findViewById(R.id.event_picture);
-        eventNameView = findViewById(R.id.event_name_view);
-        eventDetailsView = findViewById(R.id.event_details_view);
-        eventRulesView = findViewById(R.id.event_rules_view);
-        eventDeadlineView = findViewById(R.id.event_deadline_view);
-        eventCountdownView = findViewById(R.id.event_countdown_view);
-        eventPriceView = findViewById(R.id.event_price_view);
-
-        // Back button to navigate back to the previous activity
-        ImageButton eventBackButton = findViewById(R.id.event_back_button);
-        eventBackButton.setOnClickListener(new View.OnClickListener() {
-            @Override
-            public void onClick(View v) {
-                finish(); // Ends the current activity and returns to the previous screen
-            }
-        });
-
-        // Retrieve event data from intent and provide fallback values
-        Intent intent = getIntent();
-        String name = intent.getStringExtra("name");
-        String details = intent.getStringExtra("details");
-        String rules = intent.getStringExtra("rules");
-        String deadline = intent.getStringExtra("deadline");
-        String startDate = intent.getStringExtra("startDate");
-        String price = intent.getStringExtra("price");
-        String imageUriString = intent.getStringExtra("imageUri");
-
-
-        // Set data to views with null-checks
-        eventNameView.setText(name != null ? name : "No name");
-        eventDetailsView.setText(details != null ? details : "No details");
-        eventRulesView.setText(rules != null ? rules : "No rules provided");
-        eventDeadlineView.setText(deadline != null ? deadline : "No deadline");
-        eventCountdownView.setText(startDate != null ? "Starts in: " + startDate : "No start date");
-        eventPriceView.setText(price != null && !price.equals("0") ? "$" + price : "Free");
-        // Load image if URI is available
-        try {
-            if (imageUriString != null && !imageUriString.isEmpty()) {
-                Uri imageUri = Uri.parse(imageUriString);
-                eventImageView.setImageURI(imageUri); // Attempt to load directly
-            } else {
-                eventImageView.setImageResource(R.drawable.placeholder_event);
-            }
-        } catch (Exception e) {
-            e.printStackTrace();
-            eventImageView.setImageResource(R.drawable.placeholder_event); // Fallback if loading fails
-        }
-    }
 }
