--- Project_v2/app/src/main/res/layout/home_page.xml
@@ -1,56 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<androidx.constraintlayout.widget.ConstraintLayout
-    xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:app="http://schemas.android.com/apk/res-auto"
-    xmlns:tools="http://schemas.android.com/tools"
-    android:id="@+id/main"
-    android:layout_width="match_parent"
-    android:layout_height="match_parent"
-    tools:context=".MainActivity">
-
-    <!-- Header Layout -->
-    <TextView
-        android:id="@+id/header_title"
-        android:layout_width="0dp"
-        android:layout_height="wrap_content"
-        android:text="Creative List View"
-        android:textColor="@color/black"
-        android:textSize="24sp"
-        android:gravity="center"
-        android:padding="16dp"
-        android:background="@android:color/darker_gray"
-        android:textStyle="bold"
-        android:textAllCaps="true"
-        app:layout_constraintTop_toTopOf="parent"
-        app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintEnd_toEndOf="parent" />
-
-    <!-- ListView -->
-    <ListView
-        android:id="@+id/list_view"
-        android:layout_width="0dp"
-        android:layout_height="0dp"
-        android:divider="@color/black"
-        android:dividerHeight="1dp"
-        android:padding="16dp"
-        app:layout_constraintTop_toBottomOf="@id/header_title"
-        app:layout_constraintBottom_toTopOf="@id/create_button"
-        app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintEnd_toEndOf="parent" />
-
-    <!-- Button for creating new entries -->
-    <Button
-        android:id="@+id/create_button"
-        android:layout_width="74dp"
-        android:layout_height="65dp"
-        android:layout_margin="16dp"
-        android:backgroundTint="@color/black"
-        android:padding="16dp"
-        android:text="+"
-        android:textColor="@color/lucky_white"
-        android:textSize="18sp"
-        app:layout_constraintBottom_toBottomOf="parent"
-        app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintStart_toStartOf="parent" />
-
-</androidx.constraintlayout.widget.ConstraintLayout>
