+++ Project_v2/app/src/main/res/values/strings.xml
@@ -18,14 +18,14 @@
     <string name="profile_phone_text">+1 (123) 456-7890</string>
     <string name="event_description_text">Lorem ipsum dolor sit amet, consectetuer adipiscing elit. Aenean commodo ligula eget dolor. Aenean massa. Cum sociis natoque penatibus et magnis dis parturient montes, nascetur ridiculus mus. Donec quam felis, Lorem ipsum dolor sit amet, consectetuer adipiscing elit. Aenean commodo ligula eget dolor. Aenean massa. Cum sociis natoque penatibus et magnis dis parturient montes, nascetur ridiculus mus. Donec quam felis,</string>
     <string name="event_rules_text">Lorem ipsum dolor sit amet, consectetuer adipiscing elit. Aenean commodo ligula eget dolor. Aenean massa. Cum sociis natoque penatibus et magnis dis parturient montes, nascetur ridiculus mus. Donec quam felis,</string>
-    <string name="event_deadline_text">Sep 19, 2024</string>
+    <string name="event_deadline_text">09-18-2024</string>
     <string name="geolocation_verification_text"><u>Location Verification Required</u></string>
     <string name="event_name_text">EDM Festival</string>
     <string name="event_countdown_text">Starts in: %1$sd: %2$sh: %3$ss</string>
     <string name="event_price_text">Free</string>
-    <string name="create_event_name_text">EDM Festival</string>
-    <string name="create_event_details_text">Lorem ipsum dolor sit amet, consectetuer adipiscing elit. Aenean commodo ligula eget dolor. Aenean massa. Cum sociis natoque penatibus et magnis dis parturient montes, nascetur ridiculus mus. Donec quam felis,</string>
-    <string name="create_event_rules_text">Only 18+ allowed</string>
+    <string name="create_event_name_text">Enter Event Name</string>
+    <string name="create_event_details_text">Enter Event Details</string>
+    <string name="create_event_rules_text">Enter Event Rules</string>
     <string name="create_event_deadline_text"> </string>
     <string name="create_event_entrants_num_text"> </string>
     <string name="create_event_attendees_num_text"> </string>
